+++ packages/google_maps_flutter/google_maps_flutter_ios_shared_code/ios/google_maps_flutter_ios/Sources/google_maps_flutter_ios/GoogleMapController.swift
@@ -583,6 +583,13 @@ class MapCallHandler: NSObject, FGMMapsApi {
     controller?.interpretMapConfiguration(configuration)
   }
 
+  func setMarkerUpdateAnimationConfiguration(
+    _ configuration: FGMPlatformMarkerUpdateAnimationConfiguration,
+    error: AutoreleasingUnsafeMutablePointer<FlutterError?>
+  ) {
+    controller?.markersController.setMarkerUpdateAnimationConfiguration(configuration)
+  }
+
   func updateMarkers(
     byAdding toAdd: [FGMPlatformMarker], changing toChange: [FGMPlatformMarker],
     removing idsToRemove: [String], error: AutoreleasingUnsafeMutablePointer<FlutterError?>